Updated StageController.m
Moved stage parameter bounds checking to the StageController class to eliminate redundancy.
Imporved the organization of the methods and properties.
Changed StartTemp, RampUpRate, etc. to Dependant properties which are accessed using the currently set StageCounter property.
Improved bounds checking to correct for opposite sign of RampUpRate.
Improved bounds checking to correct for HoldTime less than zero.
Added accessor methods to get the StartTemp, RampUpRate, etc. for a given StageNumber (taken as a parameter)